README: remove confusing instructions
Unfortunately, some people have found this information in the README and get
confused regarding how to contribute to this specific layer
(https://github.com/rockchip-linux/meta-rockchip). These instructions are a
left-over from when our layer was forked to create this layer and are not
valid.

We are not the maintainers of https://github.com/rockchip-linux/meta-rockchip.
Please don't sent patches for this layer to us.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>